scripts/requirements.txt
# Google Docs Integration Requirements
markdown==3.4.1
google-api-python-client==2.81.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0 